requirements.txt
python-telegram-bot>=20.0
spotipy>=2.23.0
python-dotenv>=1.0.0
pymongo>=4.6.0
dnspython>=2.4.0
